+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,18 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
-0.5.74
+0.5.75
 ++++++
 
 * Add support for web application routing.
+
+++++++
+* Refactor: Removed redundant `--disable-workload-identity` flag. User can disable the workload identity feature by using `--enable-workload-identity False`.
+
+0.5.74
+++++++
+
+* Add command `aks trustedaccess role list`.
 
 0.5.73
 ++++++
@@ -90,7 +98,7 @@ Pending
 ++++++
 
 * Update to use 2022-03-02-preview api version.
-* Add support for `--format` parameter in `az aks get-credentials` command.  
+* Add support for `--format` parameter in `az aks get-credentials` command.
 
 0.5.60
 ++++++